Custom page for monero-ecosystem site
Hi, the copy paste from my other documentation didn't format very well so this edit should be better suited to the larger font size used on this page. It's also more concise with links to my main documentation where needed. Thanks for initially creating this page as I needed that nudge to produce this, I've been so busy with work sorry.

Added "- Node" to tags as it's the core function of the project
Content trimmed and spell checked
Removed my banner image so this is more consistent with other ecosystem pages